pyproject.toml
[project]
name = "kaggle-nvidia-nemotron"
version = "0.1.0"
description = "Fine-tuning and inference pipeline for NVIDIA Nemotron Model Reasoning Challenge"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.4.0",
    "transformers>=4.48.0",
    "datasets>=3.0.0",
    "peft>=0.14.0",
    "trl>=0.12.0",
    "vllm>=0.7.0; sys_platform == 'linux'",
    "wandb>=0.19.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "unsloth>=2026.3.8",
]

[project.optional-dependencies]
unsloth = ["unsloth>=2024.12"]
dev = ["pytest", "ipython", "jupyter"]

[dependency-groups]
dev = ["pytest", "ipython"]
